requirements.txt
aiogram==2.25.2
alembic
bitcoinrpc
flask
qrcode
python-dotenv
requests
SQLAlchemy
solana
web3
yoomoney
